+++ mayan/apps/navigation/templatetags/navigation_tags.py
@@ -21,23 +21,6 @@ def get_menu(name):
 
 
 @register.simple_tag(takes_context=True)
-def get_menu_links(context, name, source=None, sort_results=None):
-    return Menu.get(name).resolve(context=context, source=source, sort_results=sort_results)
-
-
-@register.simple_tag(takes_context=True)
-def get_menus_links(context, names, source=None, sort_results=None):
-    result = []
-
-    for name in names.split(','):
-        for links in Menu.get(name=name).resolve(context=context, sort_results=sort_results):
-            if links:
-                result.append(links)
-
-    return result
-
-
-@register.simple_tag(takes_context=True)
 def get_sort_field_querystring(context, column):
     return column.get_sort_field_querystring(context=context)
 
@@ -72,6 +55,33 @@ def get_source_columns(context, source, exclude_identifier=False, only_identifie
 def resolve_link(context, link):
     # This can be used to resolve links or menus too
     return link.resolve(context=context)
+
+
+@register.simple_tag(takes_context=True)
+def navigation_resolve_menu(context, name, source=None, sort_results=None):
+    result = []
+
+    menu = Menu.get(name)
+    link_groups = menu.resolve(context=context, source=source, sort_results=sort_results)
+
+    if link_groups:
+        result.append({'link_groups': link_groups, 'menu': menu})
+
+    return result
+
+
+@register.simple_tag(takes_context=True)
+def navigation_resolve_menus(context, names, source=None, sort_results=None):
+    result = []
+
+    for name in names.split(','):
+        menu = Menu.get(name=name)
+        link_groups = menu.resolve(context=context, sort_results=sort_results)
+
+        if link_groups:
+            result.append({'link_groups': link_groups, 'menu': menu})
+
+    return result
 
 
 @register.simple_tag(takes_context=True)